# docs/reference/openapi.yaml
openapi: 3.1.0
info:
  title: Huginn API
  version: 0.1.0
  description: |
    Initial contract scaffold for Huginn's HTTP API.
    This spec currently focuses on core session/chat, agent, workflow, connection,
    and system endpoints used by the web app. It is intended to be expanded
    incrementally as additional routes are hardened.
servers:
  - url: http://localhost:8080
    description: Local development server
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/sessions/{id}/artifacts:
    get:
      tags: [Sessions]
      summary: List session artifacts
      responses:
        "200":
          description: Session artifact summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SessionArtifactSummary"
    post:
      tags: [Sessions]
      summary: Create session artifact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionArtifactCreateRequest"
      responses:
        "201":
          description: Artifact created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtifactCreateResponse"
  /api/v1/sessions/{id}/artifacts/{artifact_id}:
    get:
      tags: [Sessions]
      summary: Get session artifact
      responses:
        "200":
          description: Artifact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionArtifact"
    put:
      tags: [Sessions]
      summary: Update session artifact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArtifactStatusUpdateRequest"
      responses:
        "200":
          description: Artifact status update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtifactStatusResponse"
    delete:
      tags: [Sessions]
      summary: Delete session artifact
      responses:
        "204":
          description: Artifact deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: HuginnToken
...
  schemas:
...
    SessionArtifactSummary:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        kind:
          type: string
        status:
          type: string
        mime_type:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, title]
      additionalProperties: true
    SessionArtifact:
      type: object
      properties:
        id:
          type: string
        session_id:
          type: string
        title:
          type: string
        kind:
          type: string
        status:
          type: string
        content:
          type: string
      required: [id]
      additionalProperties: true
    SessionArtifactCreateRequest:
      type: object
      properties:
        title:
          type: string
      required: [title]
      additionalProperties: true
    ArtifactCreateResponse:
      type: object
      properties:
        id:
          type: string
      required: [id]
    ArtifactStatusUpdateRequest:
      type: object
      properties:
        status:
          type: string
        reason:
          type: string
      required: [status]
    ArtifactStatusResponse:
      type: object
      properties:
        status:
          type: string
      required: [status]
...